src/config.py
"""
EAIP Manager 配置
服务器端工具配置
"""

import os
from pathlib import Path
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

# 项目根目录
BASE_DIR = Path(__file__).resolve().parent.parent

# 日志目录
LOGS_DIR = BASE_DIR / "logs"
LOGS_DIR.mkdir(exist_ok=True)

# 日志配置
LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
LOG_FILE = LOGS_DIR / "manager.log"

# AIPKG 配置
DEFAULT_COMPRESSION = "gzip"
DEFAULT_COMPRESSION_LEVEL = 6
DEFAULT_ENCRYPTION_ITERATIONS = 100000

# 密码策略
MIN_PASSWORD_LENGTH = 12
REQUIRE_UPPERCASE = True
REQUIRE_LOWERCASE = True
REQUIRE_DIGIT = True
REQUIRE_SPECIAL = True
