requirements.txt
# EAIP Manager - 服务器端工具依赖

# 加密
cryptography>=41.0.0

# 环境变量
python-dotenv>=1.0.0

# 进度条
tqdm>=4.66.0

# 测试（可选）
pytest>=7.4.0
